requirements.txt
Flask==3.0.0
Flask-Session==0.5.0
requests==2.31.0
charset-normalizer==3.3.2
python-docx==1.1.0
waitress==2.1.2
flask
flask-session
requests
charset_normalizer
python-docx
pypdf2
gunicorn